requirements.txt
arrow==0.17.0
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
PyJWT==1.7.1
pylint==2.6.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
rq==1.5.2
rq-dashboard==0.6.1
six==1.15.0
toml==0.10.1
twilio==6.45.4
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
